perf: Dsql connection reuse (#400)
* Reuse DSQL connections

This adds reuse to DSQL connections. The DSQL MCP server now maintains a
persistent connection and reuses it across queries.

* remove unused consts

* Fix dsql unit tests

* add session token to docker ex

* Add --profile option to dsql

Add a --profile option as well as AWS_PROFILE.

* DSQL readme updates

* Don't validate connection

This saves a round trip on execution by just trying to use the
connection and creating a new one if it fails before retrying execution.

* fix pyright

* Add coverage

---------

Co-authored-by: Alain Krok <alkrok@amazon.com>