fix(tests): Import __version__ for aws-documentation-mcp tests (#1097)
There was a hard coded value assertion in a test verifying the
functionality of fetching the version of the aws-documentation-mcp,
which required updating when the version numbers were bumped.

This change imports `__version__` in the aws-documentation-mcp
library, and uses it in test assertions, to remove any manual
interventions required for tests in the future.

Co-authored-by: Jonathan Lim <lmjonat@amazon.com>